refactor: consumer parameters form validation
